// src/navigation/types.ts
export type RootStackParamList = {
  Splash: undefined;
  Auth: undefined;
  Main: undefined;
  CoinDetail: { tokenId?: string; tokenName: string; symbol: string };
};

export type MainTabParamList = {
  Home: undefined;
  Scanner: undefined;
  Game: undefined;
  Hunt: undefined;
  Leaderboard: undefined;
  Profile: undefined;
};